+++ example-tsp/src/main/java/es/urjc/etsii/grafo/TSP/Main.java
@@ -4,6 +4,6 @@ import es.urjc.etsii.grafo.solver.Mork;
 
 public class Main {
     public static void main(String[] args) {
-        Mork.start(args);
+        Mork.start(args, false);
     }
 }