+++ template/src/main/java/es/urjc/etsii/grafo/__RNAME__/Main.java
@@ -4,6 +4,8 @@ import es.urjc.etsii.grafo.solver.Mork;
 
 public class Main {
     public static void main(String[] args) {
-        Mork.start(args);
+        // TODO Configure if maximizing or minimizing objective function.
+        // false to minimize, true to maximize
+        Mork.start(args, false);
     }
 }